requirements.txt
tensorflow
pandas
numpy
scikit-learn
matplotlib
seaborn
gradio
pillow
